Move BLE example to platform-specific/cc26xx
This commit moves the BLE/IPv6 example to the platform-specific directory, since we only support one platform currently. The pull also tidies up the example Makefile a bit: We no longer need to specify the name for project-conf.h so we remove this. We add PLATFORMS_ONLY and BOARDS_ONLY