uip-nd6: bugfix on address resolution by Neighbor Discovery protocol
nbr_table_update_lladdr() fails to update the lladdr of a nbr with a
new one when the new lladdr is used in nbr_table. This causes
communication errors when NS/NA is employed.

uip_ds6_nbr_update_ll() is introduced to resolve this issue. And
nbr_table_update_lladdr() is removed since it's not used any more.