Specify default tun/tap name on Mac OS
e5c7437 fixed a bug that was preventing multiple instances of tunslip6 from running. This same commit however broke tunslip6 on Mac OS, as documented in contiki-os/contiki#1560 as well as in contiki-ng/contiki-ng#466

Basically the commit in question merely removed some code. This commit puts that code back in, but makes it conditional for Mac OS.

Fixes #466