coap-blocking-api: Signal timeout to the callback function
Currently there is no convenient way for the caller of COAP_BLOCKING_REQUEST
to tell if the request timed out.

This change calls the request_callback with a NULL pointer as the argument
on timeout. Thus the caller will be able to handle a timeout by checking
for NULL in the callback function.

Examples are updated to handle being passed NULL.