Add new radio constant: RADIO_CONST_MAX_PAYLOAD_LEN
This new constant is mandatorily supported by all radio driver get_value() functions.

It represents the maximum payload the radio driver is able to handle.

This includes the MAC header and MAC payload, but not any tail bytes added automatically by the radio. For example, in the typical case of .15.4 operation at 2.4GHz, this will be 125 bytes (127 bytes minus the FCS / CRC16).

This is the maximum number of bytes that:
  * The MAC layer will ask the radio driver to transmit.
    This corresponds to the payload_len argument of the prepare() and
    send() and the transmit_len argument of transmit().
  * The radio driver will deliver to the MAC layer after frame reception.
    The buf_len of the read function will be at greater than or equal to
    this value. The read() function may not return a value greater than
    this constant.