Fixed bug in CC1310 prop RF settings
A bug was introduced in the CC1310 prop RF settings when fixing a bug
with BLE beaconing. In short, prop RF settings are shared between CC1310
and CC1350, and rfMode was set to RF_MODE_MULTIPLE which is only
supported on CC1350. This is now split between CC1310 and CC1350.