Add support for the .upload make target for platform simplelink
This commit allows users of platform simplelink to program their devices using the .upload make target. This will only work for those simplelink boards that can actually be programmed using the cc2538-bsl script. As of the time of writing this commit, this can only be done with the old `x0` devices (cc13x0 and cc26x0), but not the newer `x2` devices. Changes to the BSL script itself likely required before these platforms can be supported.

Closes #1036