Enable the entire CC2538 RAM for Zoul border routers
The default zoul build configuration allows the chip to enter PM2. When PM2 is allowed we only use 16 out of 32KB of the CC2538's SRAM. This is because the low 16KB do not have retention in PM2.

However, since the border router's radio is always on, the zoul will never actually drop to PM2. This commit therefore disables PM2 altogether in the firmware, allowing us to use the entire 32KB of the chip's SRAM.

Ultimately this allows us to build zoul border-routers with TSCH support.