Add support for multiple docker containers
This commit allows us to create and use a separate docker container image each time `tools/docker` gets changed. This means that `tools/docker` changes in one branch will not influence travis tests running against a different branch (potentially causing them to fail, as was the case for #1084 & #1085). From now on, each travis build will either pull and use the correct container image, or build a new one.

Each image is tagged with the hash of the most recent commit that changed `tools/docker`. Let's call this `CONTIKER_TAG`. This is retrieved by running:

```
$ git log -1 --oneline -- tools/docker/ | cut -d" " -f1
5310e7ae5
```

Travis tests will attempt to pull `contiker/contiki-ng:$CONTIKER_TAG`.

Then:
* If the pull succeeds then the PR / branch does not change `tools/docker`
* If `tools/docker` changed then the `docker pull` will fail. In this case, or if any other error occurs, we build a new docker image. This new docker image will be tagged `contiker/contiki-ng:$CONTIKER_TAG`

Lastly, if we are running on `contiki-ng/contiki-ng` and if the build was triggered by a branch, push `contiker/contiki-ng:$CONTIKER_TAG` to dockerhub.

To determine whether we are running on `contiki-ng/contiki-ng` (and not on a fork), we change the test to:

```
if [ $TRAVIS_REPO_SLUG == contiki-ng/contiki-ng ]
```

This is because `$TRAVIS_SECURE_ENV_VARS` can be true even outside `contiki-ng/contiki-ng`. It is on `g-oikonomou/contiki-ng` at least.

A pleasant side-effect is that when a pull gets merged, then the build triggered by the merge will be a lot faster: The first travis job that builds the image will push it to dockerhub; jobs starting later will then succeed at pulling it and will not have to build it all over again.