Fix out-of-bounds uIP pointer on URG segments
If UIP_URGDATA is zero (the default), the amount of urgent data is
deducted from the uip_appdata pointer and from uip_len without a size
check. This can cause out-of-bounds reads (for the former) and unsigned
integer underflow (for the latter).

To fix the vulnerability, the urgent data length is checked against
urg_len independent of the value of UIP_URGDATA and bounded
appropriately.