rpl-lite: fix out-of-bounds write warning in rpl_set_prefix_from_addr()
The previous call to memset wrote zeroes after the prefix member,
causing -Warray-bounds to issue warnings in GCC. This changes the
memset to only write zeroes to the prefix member before using memcpy to
copy the prefix from the addr argument. The remaining members of the
prefix_info struct are initialized using assignment expressions after
the memcpy.